Use ordered dictionary for ptf and dut ports in get_port() function
Python dictionary is used to store the dut ports and ptf ports selected for the subinterface test, but because there's no order in python dict, sometimes the dut port could be paired with a wrong ptf port(in the function get_ports() the dut ports is returned in dict but the ptf ports is returned in list of the dict values, and they are zipped in the caller to do iteration). So use OrderedDict instead of normal dictitonary to store the selected dut ports and ptf port in get_ports.