Stop ptf container before remove it
The `testbed-cli.sh` script supports an operation which is to restart
the ptf docker container. In our lab, we ran into issue that the test
server was stuck with below error after the ptf docker container
was removed:
`
Aug  9 04:25:58 server1 kernel: [329566.985211] watchdog: BUG: soft lockup - CPU#1 stuck for 23s! [swapper/1:0]
`
When the ptf docker is removed, we used the "force_kill". Not sure if this is too
brutal and caused issue to the server. This issue happens around once a week.
It is hard to reproduce and yet very annoying.

This change added code to stop the ptf container before remove it. Need to
observe for some time to see if this change can fix the issue.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>