Restore tacacs_server after the module tacacs/test_accounting.py running. (#6117)
Description of PR
In module tacacs/test_accounting.py, the fixture check_tacacs use the function setup_tacacs_client to delete the default tacacs server, and set the ptf mgmt ip as tacacs sever ip. But it doesn't restore this config when the module finish running. We want to keep the config in consistent before and after the testcase running, so fix it.

What is the motivation for this PR?
In module tacacs/test_accounting.py, the fixture check_tacacs use the function setup_tacacs_client to delete the default tacacs server, and set the ptf mgmt ip as tacacs sever ip. But it doesn't restore this config when the module finish running. We want to keep the config in consistent before and after the testcase running, so fix it.

How did you do it?
Get the default tacacs server and put them into a list, when the module finish running, delete the ptf mgmt ipand restore the default tacacs server ip.

How did you verify/test it?
Running the test cases in this module and compare the tacacs server ip before and after running.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>