[bugfix] skip vlan/test_vlan_ping.py (#6137)
Description of PR
In pr #5708 , we skip the test cases in tests/common/plugins/conditional_mark/tests_mark_conditions.yaml. There is a merge conflict and forget to skip vlan/test_vlan_ping.py when the asic_type is broadcom. In this pr, skip this module.

What is the motivation for this PR?
In pr #5708 , we skip the test cases in tests/common/plugins/conditional_mark/tests_mark_conditions.yaml. There is a merge conflict and forget to skip vlan/test_vlan_ping.py when the asic_type is broadcom. In this pr, skip this module.

How did you do it?
Add the condition to skip vlan/test_vlan_ping.py.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>