Restore default gnmi client_auth after the module telemetry/test_telemetry.py running.  (#6124)
Description of PR
In module telemetry/test_telemetry.py, the fixture setup_streaming_telemetry use the function setup_telemetry_forpyclient to set the config gnmi client_auth false. But it doesn't restore this config when the module finish running. We want to keep the config in consistent before and after the testcase running, so fix it.

What is the motivation for this PR?
In module telemetry/test_telemetry.py, the fixture setup_streaming_telemetry use the function setup_telemetry_forpyclient to set the config gnmi client_auth false. But it doesn't restore this config when the module finish running. We want to keep the config in consistent before and after the testcase running, so fix it.

How did you do it?
Get the default gnmi client auth, and add a function restore_telemetry_forpyclient to restore the default config after the module running.

How did you verify/test it?
Running the test cases using fixture setup_streaming_telemetry and compare the config gnmi client auth before and after running.

Signed-off-by: Yutong Zhang yutongzhang@microsoft.com