A fix for the get_target_speed issue on test_psu_fans
This is a partial fix for the test_get_target_speed issue on test_psu_fans.
Specifically, this fixes the issue where the test cannot differentiate
between devices that have PSU fans that behave as intended and those that
don't even have PSU fans.

In the case of the latter, the changes in this commit skip the test instead
of misleadingly marking them as having passed.

It does this by getting a total count of the fans in all PSUs and skipping the test
if that number is 0.

The actual problem with the test is that the platform.json for the devices where it
has been failing is incomplete. Vendors will be notified as appropriate.