Improve the timeout of PR testing to ensure logs are always collected (#6152)
What is the motivation for this PR?
Currently PR testing on t1-lag may be stuck and cancelled eventually
due to timeout. After the job is cancelled, the steps for collecting
logs were not executed too. It is difficult to troubleshoot the issue
without logs.

How did you do it?
Changes of this PR:
* Adjusted the timeout value for jobs and added timeout to the step
for running tests. Timeout of the run test step is slightly smaller
than timeout of the job. So that if the run test step is
cancelled due to timeout, rest of the steps still have time to run.
* Condition of the steps collecting logs were also changed to "always()".
* Removed the agent pool specification for some jobs. They will use the
the default agent pool.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>