Add a new 'by testbed' skip condition for the pytest conditional mark (#6187)
For now we can't skip a test on a specific testbed by the pytest conditional mark. The reason why we need that is we have some tests for transceivers that may fail on the testbeds with some specific type of transceivers when there is a bug, and we want to skip the test while the bug is being handled.
And the way to get the improvement is very easy, just need to add the testbed_name item into the basic_facts dict so that we can skip test by a specific setup or setup and topology.
For example, after the improvement we can add skip condition in test_mark_conditions.yaml like:

xxx/test_xxx.py:
  skip:
    reason: "Skip running on r-tigon-21"
    conditions:
      - "'r-tigon-21' in testbed"

xxx/test_xxx.py:
  skip:
    reason: "Skip running on r-tigon-20 when topology is t1"
    conditions:
      - "'r-tigon-20-t1' in testbed"

- What is the motivation for this PR?
Add a new 'by testbed' skip condition for the pytest conditional mark so we can skip test on a specific testbed.

- How did you do it?
Add the testbed_name into the basic_facts dict.

- How did you verify/test it?
By automation and manual.