Stop exabgp before stopping and removing PTF container (#6241)
* Stop exabgp before stopping and removing PTF container

We observed test server CPU stuck issue when "testbed-cli.sh restart-ptf"
is executed to restart the PTF docker container for a deployed testbed
topology. The stuck issue has been correlated with the stop of removing
the PTF container.

There has been some efforts trying to workaround this issue:
* https://github.com/sonic-net/sonic-mgmt/pull/6131
* https://github.com/sonic-net/sonic-mgmt/pull/6191

It turns out that these changes are not working.

This change is to stop the exabgp process running in PTF container before
stopping and removing it.

Tried "testbed-cli.sh restart-ptf" over 30 times with this change, the stuck
issue is not observed. Let's take this change and keep watching to see if
the issue happen again.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>