Add delay before run test_pcied.py
in batch run, the pmon docker will restart before daemon test.

the test_pcied will be performed after docker restart.

and the pcied did not update data into redis db since the pcied is not
ready yet.

the pcied need 1 minute for main thread start to run so we delay 60 secs
in setup

Signed-off-by: RogerX87 <RogerX87@gmail.com>