Merged PR 927: Fail the run test step for nightly if sanity check failed
Fail the run test step for nightly if sanity check failed

If sanity check failed and not to recover, we stop nightly test earlier.
Consequently, the passing rate of this nightly run is no longer accurate.
To work around this issue, we can fail the Run Test step of nightly test
if sanity check failed. Then while calculating passing rate, we can tell
the difference and skip this incomplete nightly run.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>