Remove AAA and TACPLUS in config after the module test_accounting.py running.   (#6246)
Description of PR
During the module test_accounting.py running, it sets tacacs accounting in config db, which will add the key AAA and TACPLUS in config db. This cause the inconsistent in config db before and after the test case running. In this pr, we delete the key AAA and TACPLUS in config db after the module running.

Summary:
Fixes # (#4674)

What is the motivation for this PR?
During the module test_accounting.py running, it sets tacacs accounting in config db, which will add the key AAA and TACPLUS in config db. This cause the inconsistent in config db before and after the test case running. In this pr, we delete the key AAA and TACPLUS in config db after the module running.

How did you do it?
Use configlet to delete the config after test case running.

How did you verify/test it?
Check the config db before and after the test case running.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>