Shutdown Ethernet0 after the module test_neighbor_mac.py running. (#6249)
Description of PR
During the module test_neighbor_mac.py running, it start up Ethernet0. This cause the inconsistent in config db before and after the test case running. In this pr, we shutdown the Ethernet0 after the module running.

What is the motivation for this PR?
During the module test_neighbor_mac.py running, it start up Ethernet0. This cause the inconsistent in config db before and after the test case running. In this pr, we shutdown the Ethernet0 after the module running.

How did you do it?
Shutdown Ethernet0 after the module running.

How did you verify/test it?
Check the config db before and after the module running.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>