Remove RADIUS and TACPLUS in config after the module test_techsupport_no_secret.py running. (#6248)
Description of PR
During the module test_techsupport_no_secret.py running, it sets tacacs/radius password, which will add the key RADIUS and TACPLUS in config db. This cause the inconsistent in config db before and after the test case running. In this pr, we delete the key RADIUS and TACPLUS in config db after the module running.

What is the motivation for this PR?
During the module test_techsupport_no_secret.py running, it sets tacacs/radius password, which will add the key RADIUS and TACPLUS in config db. This cause the inconsistent in config db before and after the test case running. In this pr, we delete the key RADIUS and TACPLUS in config db after the module running.

How did you do it?
Use configlet to delete the config after test case running.

How did you verify/test it?
Check the config db before and after the test case running.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>