Check dut health status after running one test case. (#5817)
Description of PR
Dut health status may be changed after running one test case, such as core dump and configuration. And these changes may affect the next running test case. So we define a module level fixture to check the dut health status after running one test case, if they are different, we will reload the configuration before the module running and give a warning in the log. The pattern of the log is Check Dut health failed:

What is the motivation for this PR?
Dut health status may be changed after running one test case, such as core dump and configuration. And these changes may affect the next running test case. So we define a module level fixture to check the dut health status after running one test case, if they are different, we will reload the configuration before the module running and give a warning in the log. The pattern of the log is Check Dut health failed:

How did you do it?
Compare the running config with golden config, if they are same, continue. Otherwise, reload the testbed configuration using the config before the module running.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>