Merged PR 928: Ignore -i image url argument when IMAGE_URL is empty
If `IMAGE_URL` is not set, it's emtpy, the arguments for report_uploader.py look like this, it will take `../../results` as the value for argument `-i`, then recognize `SonicTestData` as path, that's why the script asked for database argument.

In this change, will not add -i when IMAGE_URL is empty to avoid this kind of issue.

```
+ python3 report_uploader.py -c test_result -e vmsvc1-t1-n3164-acs-1.201911#140222 -t vmsvc1-t1-n3164-acs-1 -i ../../results SonicTestData
usage: report_uploader.py [-h] [--external_id EXTERNAL_ID] [--json]
                          [--category CATEGORY] [--testbed TESTBED]
                          [--image_url IMAGE_URL | --version VERSION]
                          path [path ...] database
report_uploader.py: error: the following arguments are required: database
```

Signed-off-by: Zhaohui Sun <zhaohuisun@microsoft.com>