Merged PR 935: Capture the RC of run_tests.sh and exit early for sanity check failure
Capture the RC of run_tests.sh and exit early for sanity check failure

Previously PR #927 tried to set the Run Tests step to failed if sanity
check failed and need to exit early. However that change has bug
with checking the return code of run_tests.sh. It uses `$?` to get
return code of run_tests.sh. However, in the `elif` clause, value of
`$?` is no longer return code of run_tests.sh. It now holds value of
the previous `if` clause.

This change captured the return code of run_tests.sh in a variable
and then check that variable to decide how to exit the Run Test
step.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>