[dualtor] Fix `test_orchagent_slb` (#6272)
Approach
What is the motivation for this PR?
test_orchagent_slb fails because of the dscp check failure in the
tunnel_traffic_monitor, which is introduced by the PCBB change. So
with PCBB on, the outer packet DSCP differs from the inner one.

Signed-off-by: Longxiang Lyu lolv@microsoft.com

How did you do it?
Let's skip the dscp check for test_orchagent_slb.

How did you verify/test it?