Add support of cache issue status for conditional mark (#6313)
What is the motivation for this PR?
The conditional mark may need to check Github or other issue tracking website to get the status of issue and then decide whether to skip a test.

The problem is that when there is issue accessing the remote issue tracking website, it takes some time for the HTTP request to timeout or stop retrying. If the skipping of lots of test cases is dependent on querying issue status, the overall test time is obviously increased unnecessarily.

How did you do it?
This change added support of caching the issue status, no matter the issue status querying is success or not, the result is always cached for later use. Then in a pytest session, we only need to query any issue only once regardless how many test cases are dependent on status of the issue.

This change also improved the run_tests.sh script to clear pytest cache before running the first script.

How did you verify/test it?
Verified using some dummy test scripts.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>