Fix issue #3119918: [Functional] [QoS] | PFC / PFC Xon test keeps failing on r-tigon-11 (#6153)
Fix the PFC xon and xoff test failure on tigon-11 which is using Mellanox-SN4600c-C64

Root cause:
1. Recently, the buffer descriptor pool size has been adjusted: the total size is smaller: 204800 => 180608 and ALPHA INFINITE => 8, which means we have a smaller number of descriptors to use.
2. the test logic: send the sufficient number of packets to occupy the shared buffer proportion and then the headroom. during this time, it does not expect the packets to be dropped. however, due to a smaller descriptor pool, it will run out in advance and fail the test
3. now the failure occurs on r-tigon-11 only because it is of Mellanox-SN4600c-C64 which has the largest data buffer cells

solution: using large packets in the test, like 300
By using large packets, we can send less number of packets to occupy all the data buffer cells. We encountered a similar issue in all watermark tests. then we use larger packets for testing to resolve the issue.
Changing the packet size does not suffice. The number of cells to occupy should be also adjusted. Check watermark tests for details.

- What is the motivation for this PR?
Fix the qos test failed on Mellanox-SN4600c-C64 hwsku

- How did you do it?
Modify the test to use large pkt to do the test, and adjust the cell occupancy to make the test pass

- How did you verify/test it?
Run the test case , and it pass