Split Running job step in testbedV2 pipeline.  (#6338)
Description of PR
In testbed v2, we only have one Running test step to trace all status of a testplan. In this pr, we split this Running test into Lock testbed, Prepare testbed and Running test to trace the status more explicitly.

What is the motivation for this PR?
In testbed v2, we only have one Running test step to trace all status of a testplan. In this pr, we split this Running test into Lock testbed, Prepare testbed and Running test to trace the status more explicitly.

How did you do it?
Split Running test step into three steps.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>