[image 20191130] [masic] skip power off reboot in 20191130 image for masic (#6435)
What is the motivation for this PR?
test_reboot_power_off was skipped in 20191130.78 as no PSU controller
provided in this version.

Starting 20191130.80 PSU controller is there, so test is not skipped.

After a power off reboot, we want to ensure all critical services are up.
When checking reboot-cause, the output of ‘show reboot-cause’ performed
on dut is ‘Unknown’, which is not as expected ‘Power off’ and caused
step check reboot cause to fail.

The reason why ‘show reboot-cause’ fail to get reboot-cause is that:
• systemd failed to find hardware reboot cause due to:
o Sep 22 03:53:41.916464 str-n3164-acs-2 WARNING process-reboot-cause:
sonic_platform package not installed. Unable to detect hardware reboot causes.

• Then it tried to get software reboot cause by reading from file
/host/reboot-cause/reboot-cause.txt and gives ‘Unknown’

According to following code, vendor should provide sonic_platform
package for this test to pass:

def find_hardware_reboot_cause():
    hardware_reboot_cause = None

    # Until all platform vendors have provided sonic_platform packages,
    # if there is no sonic_platform package installed, we only provide
    # software-related reboot causes.
    try:
        import sonic_platform

        platform = sonic_platform.platform.Platform()