sonic-mgmt: Fix test_show_platform_syseeprom on Arista devices
'show platform syseeprom' is vendor specific and on Arista
devices this information is the contents of the prefdl.

Change to minimally validate that the output of the show
command is non empty on an Arista device as the output
does not conform to the 'syseeprom_info' format.
Use chassis get_system_eeprom_info api to have
'syseeprom_info' formatted data on an Arista device.

Signed-off-by: Aaron Payment <aaronp@arista.com>