Correct the qos test_buffer.py for innovium platform (#6454)
When we used the innovium platform to run the qos/test_buffer.py. we encounter error.

Signed-off-by: RogerX87 <RogerX87@gmail.com>

What is the motivation for this PR?
To correct the error and run this case successfully.

How did you do it?
correct the array index. it's 3-d array now, not 2-d array as before.

How did you verify/test it?
Run the test with wistron Innovium platform.

Any platform specific information?
Innovium platform only.