fix testbedprocessing hostvar generation
sonic-mgmt: Fix TestbedProcessing of host vars file name

Ansible files in ansible/host_vars/* are named to match
the host that the vars are to be loaded for. The filename
is case-sensitive to the host names that are in the
inventory files. Need to remove the forcing of the file
name to be upper case.

Signed-off-by: Aaron Payment <aaronp@arista.com>