Remove autopep8 from pre-commit check
The function of autopep8 is duplicated with flake8 and it
updates scripts in place by default. The other problem is
that autopep8 does not show what it has changed in command
output. It simply raises a "failed" flag. We have to run
pre-commit check locally to figure out what's the problem.

Considering these drawbacks, we can remove autopep8 from
pre-commit check. Just use flake8 check should be enough.

This change removed the autopep8 pre-commit check from
the `.pre-commit-config.yaml` config file.

The other thing is that pre-commit only supports python3.
This change improved the github comment template for pre-commit.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>