Limit the length of pre-commit check results
PR 6523 introduced an issue and caused the pre-commit
check results cannot be rendered in Github comment.

What's more, the pre-commit check results could be too
long to be displayed in a Github comment.

This change fixed the rendering issue and also added code
to limit the length of pre-commit check results.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>