[sub_port] [backend] Update acl template based on the test ports (#6556)
Signed-off-by: Neetha John <nejo@microsoft.com>

What is the motivation for this PR?
For storage backend, new backend acl was added (sonic-net/sonic-utilities#2236). This caused regression for sub port testcases running on 't0-backend' topology. Modified the testcase to update the acl template based on the selected test ports

How did you verify/test it?
Ran the sub port testcases with the change and they passed