everflow test case fix (#6561)
What is the motivation for this PR?
This PR #6225 introduced one issue for T0/T1. The pytest will be crashed when running everflow test.

The reason is because the implementation of one fixture, setup_recycle_port, which include the yield but not run into the yield for non t2 case, which will hit one known issue of pytest, which is fixed in newer version of pytest but not in what we are using now.
pytest-dev/pytest#7061

How did you do it?
Change the condition to make sure the yield would be executed

How did you verify/test it?
Run everflow test.

Any platform specific information?
Supported testbed topology if it's a new test case?