Skip test_cacl_application to unblock sub module update of sonic-host-services (#6585)
What is the motivation for this PR?
Recent change in caclmgrd caused the expected ACL list to change.
test_cacl_application fails because the expected ACLs are not present on DUT.
This is occurring with latest sonic-host-services changes, and is blocking sub module update.
Test case has to be updated, but updated test case will fail without the latest changes from sonic-host-services. This PR is to ensure that sub module is updated so that the test case can pass.
issue: sonic-net/sonic-buildimage#12464
submodule update PR that is blocked: sonic-net/sonic-buildimage#12195

How did you do it?
skip test_cacl_application.