Add start bgp step in bgp bounce test (#6566)
In the bgp bounce test case, template is used to update the bgp config. And it is required to have 'admin_status' in the BGP_NEIGHBOR of configDB. Otherwise, an error will be thrown, and the peer may not be updated:

ERR bgp#bgpcfgd: Peer '(default|fc00::5e)': Can't update the peer. Only 'admin_status' attribute is supported
The initial config of community setups doesn't contain 'admin_status', so need to configure it before the test.

There is a similar issue and the error was ignored:
#5572
But in this test just ignoring the error may cause false positive result.