Skip "FLEX_COUNTER_DELAY_STATUS" when checking running config.  (#6624)
Description of PR
In pr #6541 , it changed the logic of generating golden config.It will cause more reload during test running, and add test running time. Most reload is caused by the "FLEX_COUNTER_DELAY_STATUS" in "FLEX_COUNTER_TABLE". In this pr, we temporarily skip this key, and we fix the issue, we will add this check back.

What is the motivation for this PR?
In pr #6541 , it changed the logic of generating golden config.It will cause more reload during test running, and add test running time. Most reload is caused by the "FLEX_COUNTER_DELAY_STATUS" in "FLEX_COUNTER_TABLE". In this pr, we temporarily skip this key, and we fix the issue, we will add this check back.

How did you do it?
Skip check the key "FLEX_COUNTER_DELAY_STATUS" when checking the running config.

Signed-off-by: Yutong Zhang <yutongzhang@microsoft.com>