[devutils] Support printing the logs of `ansible/devutils` to `STDOUT` (#6622)
What is the motivation for this PR?
* Support printing logs of `ansible/devutils` to `STDOUT`
* Update the level of some logs in `snmp_pdu_controller`
* Fix some errors reported by flake8.

How did you do it?
I added an optional parameter `--log-level` to `ansible/devutils`. By default, it's set to close and will not print logs to `STDOUT`, which is the same as current behavior. If you set a log level, then the logs filtered out will be printed to `STDOUT`.
I also updated the level of some logs in `snmp_pdu_controller` and fixed some errors reported by flake8.

How did you verify/test it?
Verified on physical testbeds.

Signed-off-by: Zhijian Li <zhijianli@microsoft.com>