[pre-commit] Fix style issues in files under the docs folder
pre-commit is static analysis tool introduced recently. This tool
is not able to do diff-only check. It checks the whole files touched
by PR. We can't blame PR author for legacy issues. That's why
currently the pre-commit check is only optional.

To ensure that we can make pre-commit a mandatory check for PRs
submitted to this repository, we need to fix all the legacy issues
complained by pre-commit.

This change fixed all pre-commit issues in files under the docs folder.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>