[pre-commit] Fix style issues in `tests/conftest.py`
pre-commit is a static analysis tool introduced recently. This tool is not able to do diff-only check. It checks the whole files touched by PR. We can't blame PR author for legacy issues. That's why currently the pre-commit check is only optional.

To ensure that we can make pre-commit a mandatory check for PRs submitted to this repository, we need to fix all the legacy issues complained by pre-commit.

This change fixes the style issues in `tests/conftest.py`.

What's more, built-in function `reduce` is used in `tests/conftest.py`.
This built-in function is no longer supported in python3. This change
updated the code using `reduce` to use `sum` instead.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>