[dual_tor_io] Improve && fix `dual_tor_io` (#6642)
Approach
What is the motivation for this PR?
Fix that nightly is stuck in test_link_drop.py

Signed-off-by: Longxiang Lyu lolv@microsoft.com

How did you do it?
The root cause to this issue is that the sniffer thread DualtorIO.scapy_sniff, though the sender kills the sniffer process in ptf, the following Ansible command doesn't get returned.
This PR introduces few fixes:
1. Use supervisord to control the ptf sniffer start/stop instead of letting the sniffer thread blocking waiting the ptf sniffer to finish.
2. Let the sniffer thread to wait on a Threading.Event object, which could be set if the test wish to stop the sniffer early.
Improve the the class DualToRIO by removing unneccesary threads.

The rationale chooses using supervisord over using nohup to run ptf sniffer in the backgroud is that, with ptf sniffer running in the background with nohup, SIGINT will be ignored, sending SIGINT to the sniffer won't not stop the sniffer. For some of the dualtor-io testcases, the sender wishes to stop the sniffer early by calling stop_sniffer_early(). So stop_sniffer_early() could not stop the ptf sniffer in this case, the ptf sniffer keeps running and no pcap file generated, the testcase will fetch the wrong pcap(the pcap from the last dualtor-io testcase) and throw failures.
So the key is to ensure that the ptf sniffer could receives the SIGINT and get stopped gracefully(with captured files saved to files). With supervisord, this could be guaranteed by using SIGINT as stop signal for the ptf sniffer process.
Now with supervisorctl to run ptf sniffer, there is no need to run the sender/sniffer in threads. And the I/O ops could be achieved with sequential function calls:
```
    def send_and_sniff(self):
        """Start the I/O sender/sniffer."""
        try:
            self.start_sniffer()
            self.send_packets()
            self.stop_sniffer()
        except Exception:
            self.force_stop_ptf_sniffer()
            raise

        self.fetch_captured_packets()
```
How did you verify/test it?
Run test_link_drop.py on dualtor-mixed testbed.