[pre-commit] Fix style issues in test scripts under `tests/arp` folder (#6694)
* [pre-commit] Fix style issues in test scripts under 'tests/arp' folder

What is the motivation for this PR?
pre-commit is a static analysis tool introduced recently. This tool is
not able to do diff-only check. It checks the whole files touched by PR.
We can't blame PR author for legacy issues. That's why currently the
pre-commit check is only optional.

To ensure that we can make pre-commit a mandatory check for PRs
submitted to this repository, we need to fix all the legacy issues
complained by pre-commit.

How did you do it?
This change fixes the style issues of test scripts under the 'tests/arp'
folder

How did you verify/test it?
Any platform specific information?
Supported testbed topology if it's a new test case?