Added a new logic to reboot logs print. (#6334)
What is the motivation for this PR?
In the current situation, a relevant information for debug wasn't printed in the log in case of failure, only the specific fail reason.
In addition, in case the test passes, there were no stats about the data & control plane downtime during the reboot.
These stats are important to measure the performance of the reboot process and to detects degradations in downtime.

My fix prints important debug info from the reboot log in case of failure, and a summary of downtime in case of success.
By that, only relevant information will appear in log, and will make the log much more readable and useful.

How did you do it?
I created a method that reads the reboot log, extracts the summary section using regex pattern, detects if there are failures in the test, and prints the summary if the test passed. In case there are failures, the all reboot stats will be printed to the log for debug purposes.

How did you verify/test it?
I ran full regression on community tests.