[test]: Ignore MUX_LINKMGR for DUT health check (#6689)
Add "MUX_LINKMGR" to the set of tables excluded from the check, and exclude it from pre-test and post-test table checks as well

Currently, the DUT health check diffs CONFIG_DB tables before and after test sessions. However, the run_icmp_responder fixture changes the MUX_LINKMGR table in CONFIG_DB and does not clean up after itself, which is expected. The change is to increase the linkmgrd heartbeat interval (introduced here #4669). It's appropriate to persist this change because the ICMP responder and mux simulator may not have the same performance as a real server and mux cable.

Signed-off-by: Lawrence Lee <lawlee@microsoft.com>