Merged PR 1266: Resolve test_cold_reboot failure for 3800
Merge resolving test_cold_reboot failure for 3800

**What**
16196968 - [SONIC-NIGHTLY]platform_tests.test_reboot.test_cold_reboot failed to fix the issue for 3800 Testbed
test_cold_reboot testcase is failing for SN3800 with the below error
AssertionError: Not all transceivers are detected or interfaces are up in 300 seconds

AssertionError: Status is not expected, output [u'Ethernet252', u'Not', u'present']

**Root cause analysis**
It seems that the connection graph (starlab_connection_graph.xml) for vms12-t0-3800 has an entry for Ethernet252 with a transceiver connected to str-7260cx3-64-17 port Ethernet32/1.
However, the same peer port is also connected to str-msn4600c-acs-02 port Ethernet116.
Since vms12-t0-3800 Ethernet252 has no transceiver connected, we need to update the xml file to remove the info for vms12-t0-3800 Ethernet252.

**Summary of fix**
Updated the file to remove Ethernet252 mapping for vms12-t0-3800

**Unit test**
1. Verified test_check_sfp_presence is passing
2. Verified platform_tests.test_reboot.test_cold_reboot is passing