[autorestart]: Enable autorestart for all features before test begins (#6800)
What is the motivation for this PR?
If we run test_pretest.py and then run test_container_autorestart.py, autorestart test will fail for teamd/swss/syncd containers.
The reason for this is, test_pretest disables autorestart for all features.
when the test begins running for a specific container, it enables autorestart for that feature and expects that feature to restart.
When the test starts for teamd, it disables teamd autorestart. When teamd goes down, swss /syncd also goes down.
Teamd does not start back up because swss/sycnd are down and they do not come up because the autorestart flag is disabled for swss/syncd by pretest.
This causes test to fail for teamd docker, test fails for swss/syncd dockers because they do not come up after teamd goes down.
How did you do it?
Enable autorestart for all features before the test starts, that way swss/syncd autorestart will be enabled.

How did you verify/test it?
Verified on single asic VS and chassis.