[chassis-packet]: Update linecard minigraph to remove neigh asic information to fabric asics that are not present in supervisor (#6236)
What is the motivation for this PR?
Sanity tests on packet chassis fails because the admin status and oper status do not match on the links that are connected to fabric asics that are not present.
Supervisor can have max number of asics and asics that are present are included in asics_present field in inventory file.
Minigraph that is generated for supervisor and Linecard will include information of asics that are not present as well.
This PR is diff is to ensure that Linecard minigraph will include neigh asic information only of fabric asics that is present in supervisor.

How did you do it?
Find the supervisor dut name for the chassis testbed for which linecard is to be generated. If there is no supervisor in testbed, default value of None will be used.
Get the asics_present list of supervisor dut
Pass the values of card_type and asics_present to topo_facts library
Based on card_type, if card_type is not supervisor and we have the list of asics that are present, then only keep the asics that are present in linecard minigraph.
How did you verify/test it?
on packet chassis linecard, verified that the admin status on interfaces to fabric asics that are not present are down.
on packet chassis supervisor - no change in minigraph, the swss/sycnd dockers will not be started on asics that are not present so admin status and oper status mismatch cannot occur.
on voq chassis - no change in any minigraph
on multi-asic device (non chassis) - no change in minigraph