[snmp]: Fix test_snmp_phy_entity psu test (#6802)
test_snmp_phy_entity is failing on chassis due to the below reasons:

psu test case run on linecard, but linecard does not have PSU_INFO in state_db
Test failing on SUP because test case turns off one of the pdu outlet, the loop to check off psu on the sup finds a psu that is not present on sup. This is because only status of psu is checked, but status should be checked only if psu is present.
How did you do it?
Skip psu test if PSU_INFO is not found in state_db
Check for presence and out of the PSUs that are present, look for the PSU that is turned off.
How did you verify/test it?
Tested on chassis and single asic device.